backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-anthropic
langchain-core
httpx
python-dotenv
pydantic
langgraph-cli[inmem]
langsmith